fix: dead code/quality, remove stats.json
